\name{mu}
\alias{mu}
\title{Compute the mean of Nakagami distribution}
\description{Compute the mean of Nakagami distribution}
\usage{mu(x, y)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ The shape parameter}
  \item{y}{ The scale parametr}
}
\value{a numeic value represents the mean of Nakagami distribution}
\author{Riyadh Al-Mosawi}
\examples{mu(1,1)}
