\name{sig}
\alias{sig}
\title{
Compute the standard devaition of Nakagami distribution
}
\description{
Compute the standard deviation of Nakagami disribution
}
\usage{
sig(x, y)
}
\arguments{
  \item{x}{The shape parameter}
  \item{y}{The scale parameter}
}
\value{
a numeic value represents the standrad deviation of Nakagami distribution
}
\author{
Riyadh Al-Mosawi
}
\examples{
sig(1,1)
}
